Working on major refactor
Implementing reducer and context hooks
