+++ package.json
@@ -36,6 +36,6 @@
     ]
   },
   "devDependencies": {
-    "eslint-plugin-react-hooks": "^2.5.0"
+    "eslint-plugin-react-hooks": "^2.5.1"
   }
 }
