+++ src/components/Form/form.component.jsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { Input, Button, Form, Row, Col } from 'antd';
 import { TagsOutlined } from '@ant-design/icons'; // Icon
 
-import Kalender from '../Calendar/calendar.component';
+import Calendar from '../Calendar/calendar.component';
 
 // Custom Hooks
 import useForm from '../../utils/useForm';
 import useTodos from '../../utils/useTodos.js';
 
 // Functions
-import onFormSubmit from '../../functions/onFormSubmit.js';
+import {FormSubmit} from './FormSubmit';
 
 // We destructure the props sent from the parent
 // We wont need these after the refactoring
@@ -27,7 +27,8 @@ const TodoForm = ({
   const [todos, handleTodos] = useTodos(); //TODO
 
   return (
-    <Form onFinish={onFormSubmit}>
+    <>
+    <Form onFinish={FormSubmit}>
       <h3>
         <b>Add TODO item</b>
       </h3>
@@ -49,7 +50,7 @@ const TodoForm = ({
 
         <Col>
           <Form.Item name="calendar" label="Date picker">
-            <Kalender onChange={setDate} />
+            <Calendar onChange={setDate} />
           </Form.Item>
         </Col>
       </Row>
@@ -60,6 +61,9 @@ const TodoForm = ({
         </Button>
       </Row>
     </Form>
+    
+    <Button type="primary" onClick={() => console.log(form)}>Console.log useForm value </Button>
+    </>
   );
 };
