+++ src/constants/FORM_COLUMNS.js
@@ -1,0 +1,19 @@
+export const FORM_COLUMNS = [
+  {
+    title: 'Title',
+    dataIndex: 'title',
+    key: 'title'
+  },
+
+  {
+    title: 'Date',
+    dataIndex: 'date',
+    key: 'date'
+  },
+
+  {
+    title: 'Action',
+    key: 'action',
+    dataIndex: 'action'
+  }
+];
