+++ src/utils/useForm.js
@@ -8,11 +8,14 @@ const useForm = () => {
   const handleChange = e => {
       e.persist();
       setState(state => ({ ...state, [e.target.name]: e.target.value }))
-      console.log(state);
+      //console.log(state);
   
   }
 
   return [state, handleChange];
 };
+
+
+
 
 export default useForm;
